aplikasikan hak akses

+++ antrian_gelas_ari_ek_delete.php
@@ -1,5 +1,25 @@
-<?php
-include('session_manager.php');
+
+<?php 
+
+include('session_manager.php'); 
+checkRoleAccess([
+    'Superadmin', 
+    'Kabag', 
+    'Kasie', 
+    'Kasubsie', 
+    'Admin QC', 
+    'Koordinator QC', 
+    // 'Mandor Off Farm', 
+    // 'Analis Off Farm', 
+    'Mandor On Farm', 
+    'Analis On Farm', 
+    // 'Operator Pabrikasi',
+    // 'Staff Teknik',
+    // 'Staff Tanaman',
+    // 'Staff TUK',
+    // 'Direksi',
+    // 'Tamu',
+    ]);
 
 $stmt = $conn->prepare("DELETE FROM kartu_aris WHERE id=?");
 $stmt->bind_param("i", $_GET['id']);
